// src/Registro/types/typeRegistro.ts
export interface PropsRegitros {
  id: string
  date: string
  time: string
  provider: string
  remission: string
  volume: number
  realVolume: number
  fat: number
  protein: number
  temperature: number
  pH: number
  density: number
  quality: string
  user: string
  price: number
  firma: string
  notes: string[]
}

// export interface IngresoForm {
//   id: string,
//   date: string
//   time: string
//   provider: string
//   remision: string
//   volumenRemision: number
//   volumenRecibido: number
//   fat: number
//   protein: number
//   temperature: number
//   ph: number
//   density: number
//   price: number
//   quality: string
//   firma: string
//   note: string[]
// }